Unify timeline unit to millisecond
The studio and runtime had mixed timing units (seconds and milliseconds)
. This commit unifies the timings to milliseconds. The values are
however kept saved in the UIP file in seconds for backward compatibility
and also cause it is more compact (ex: 1 vs 1000).

+++ src/dm/systems/Qt3DSDMSlideCore.h
@@ -76,9 +76,9 @@ public:
      */
     virtual void GetSlides(TSlideHandleList &outSlides) const = 0;
 
-    virtual float GetSlideTime(Qt3DSDMSlideHandle inSlide) const = 0;
+    virtual long GetSlideTime(Qt3DSDMSlideHandle inSlide) const = 0;
 
-    virtual void SetSlideTime(Qt3DSDMSlideHandle inSlide, float inNewTime) = 0;
+    virtual void SetSlideTime(Qt3DSDMSlideHandle inSlide, long inNewTime) = 0;
 
     //===============================================================
     // Slide derivation
